updating graph construction code
