--- graph_construction/configs/heptrkx_plus.yaml
@@ -1,16 +1,0 @@
-# Input/output configuration
-input_dir: /tigress/jdezoort/train_1
-output_dir:  /scratch/gpfs/jdezoort/hitgraphs/heptrkx_plus_1p9
-n_files: 1770
-
-# Graph building configuration
-selection:
-    pt_min: 1.9 # GeV
-    phi_slope_max: 0.000763 
-    z0_max: 25000 
-    n_phi_sections: 1
-    n_eta_sections: 1
-    eta_range: [-5, 5]
-    endcaps: true
-    phi_reflect: false
-    remove_intersecting_edges: true
