+++ graph_construction/configs/pre-clustering.yaml
@@ -1,0 +1,41 @@
+# Input/output configuration
+input_dir: /tigress/jdezoort/train_1
+output_dir: ../../hitgraphs_1/pre-clustering_1
+n_files: 1770
+
+# Graph building configuration
+selection:
+    pt_min: 1 #GeV 
+    phi_slope_max: 0.007
+    z0_max: 350
+    n_phi_sections: 1
+    n_eta_sections: 1
+    eta_range: [-5, 5]
+    endcaps: true
+    phi_reflect: false
+    remove_intersecting_edges: true
+
+
+
+###########################################
+## RECOMMENDED GRAPH CONSTRUCTION PARAMS ##
+###########################################
+# -------------------------------------
+#  pt [GeV]  phi_slope [1/mm]  z0 [mm]
+# -------------------------------------
+## 2         0.006             275
+## 1.9       0.006             275
+## 1.8       0.006             275
+## 1.7       0.006             275
+## 1.6       0.006             275
+## 1.5       0.006             275
+## 1.4       0.0062            290
+## 1.3       0.0064            305
+## 1.2       0.0066            320
+## 1.1       0.0068            335
+## 1         0.007             350
+## 0.9       0.0072            375
+## 0.8       0.0080            425
+## 0.7       0.0085            500
+## 0.6       0.0085            500
+## 0.5       0.0085            500
